==> 2_train_test_split.ipynb <==
# 2.1. Download cleaned dataset

import os
import pandas as pd

import wandb
from utils.wandb_utils import wandb_log_artifact, get_wandb_artifact

run = wandb.init(project='ex_census_wandb', job_type='data split')

_, file_path = get_wandb_artifact(run,
                               artifact_name="census_cleaned", 
                               file_name="census_cleaned.csv", 
                               tag="latest")

print(file_path)

data = pd.read_csv(file_path, sep='\t', encoding='utf-8')
data

# 2.2. Train / Test Split

artifacts to log

- train set
- test set

from sklearn.model_selection import train_test_split

# train / test split

test_size = 0.2
seed = 42
train, test = train_test_split(data, test_size=test_size, random_state=seed)

file_path = './dataset/census_{}.csv'

train_file_path = file_path.format('train')
test_file_path = file_path.format('test')

train.to_csv(train_file_path, sep='\t', encoding='utf-8', index=False)
test.to_csv(test_file_path, sep='\t', encoding='utf-8', index=False)

metadata = {
    "test_size" : test_size,
    "random_state" : seed
}

wandb_log_artifact(run, "census_split", 
                       description='splitted train/test',
                       file_path=[train_file_path, test_file_path],
                       artifact_type = "data",
                       metadata = metadata,
                       remove_logged_file=True)

run.finish()
